requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.104.0
mysql-connector==2.2.9
pydantic==2.4.2
python-dotenv==0.21.0
python-telegram-bot==20.3
simpleaudio==1.0.4
SQLAlchemy==2.0.22
